Fix aggregations release macro
