+++ src/macros.rs
@@ -23,180 +23,6 @@ macro_rules! test_serialization {
     };
 }
 
-/// Builds query enum from given variants
-#[doc(hidden)]
-#[macro_export]
-#[cfg(not(debug_assertions))]
-macro_rules! query {
-    ($name:ident { $($variant:ident($query:ty)),+ $(,)? }) => {
-        use crate::util::*;
-
-        /// A container enum for supported Elasticsearch query types
-        #[derive(Debug, Clone, PartialEq, Serialize)]
-        #[serde(untagged)]
-        #[allow(missing_docs)]
-        pub enum $name {
-            $(
-                $variant($query),
-            )*
-        }
-
-        $(
-            impl From<$query> for $name {
-                fn from(q: $query) -> Self {
-                    $name::$variant(q)
-                }
-            }
-        )+
-
-        $(
-            impl From<$query> for Option<$name> {
-                fn from(q: $query) -> Self {
-                    if q.should_skip() {
-                        None
-                    } else {
-                        Some($name::$variant(q))
-                    }
-                }
-            }
-        )+
-
-        impl ShouldSkip for $name {
-            fn should_skip(&self) -> bool {
-                match self {
-                    $(
-                        $name::$variant(q) => q.should_skip(),
-                    )+
-                }
-            }
-        }
-    };
-}
-
-/// Builds query enum from given variants
-#[doc(hidden)]
-#[macro_export]
-#[cfg(debug_assertions)]
-macro_rules! query {
-    ($name:ident { $($variant:ident($query:ty)),+ $(,)? }) => {
-        use crate::util::*;
-
-        /// A container enum for supported Elasticsearch query types
-        #[derive(Debug, Clone, PartialEq, Serialize)]
-        #[serde(untagged)]
-        #[allow(missing_docs)]
-        pub enum $name {
-            $(
-                $variant(Box<$query>),
-            )*
-        }
-
-        $(
-            impl From<$query> for $name {
-                fn from(q: $query) -> Self {
-                    $name::$variant(Box::new(q))
-                }
-            }
-        )+
-
-        $(
-            impl From<$query> for Option<$name> {
-                fn from(q: $query) -> Self {
-                    if q.should_skip() {
-                        None
-                    } else {
-                        Some($name::$variant(Box::new(q)))
-                    }
-                }
-            }
-        )+
-
-        impl ShouldSkip for $name {
-            fn should_skip(&self) -> bool {
-                match self {
-                    $(
-                        $name::$variant(q) => q.should_skip(),
-                    )+
-                }
-            }
-        }
-    };
-}
-
-/// Builds query enum from given variants
-#[doc(hidden)]
-#[macro_export]
-#[cfg(not(debug_assertions))]
-macro_rules! query {
-    ($name:ident { $($variant:ident($query:ty)),+ $(,)? }) => {
-        /// A container enum for supported Elasticsearch query types
-        #[derive(Debug, Clone, PartialEq, Serialize)]
-        #[serde(untagged)]
-        #[allow(missing_docs)]
-        pub enum $name {
-            $(
-                $variant($query),
-            )*
-        }
-
-        $(
-            impl From<$query> for $name {
-                fn from(q: $query) -> Self {
-                    $name::$variant(q)
-                }
-            }
-        )+
-
-        impl $name {
-            /// Gets aggregation name
-            pub fn name(&self) -> String {
-                match self {
-                    $(
-                        Self::$variant(a) => a.name.clone(),
-                    )+
-                }
-            }
-        }
-    };
-}
-
-/// Builds aggregation enum from given variants
-#[doc(hidden)]
-#[macro_export]
-#[cfg(debug_assertions)]
-macro_rules! aggregation {
-    ($name:ident { $($variant:ident($query:ty)),+ $(,)? }) => {
-        /// A container enum for supported Elasticsearch query types
-        #[derive(Debug, Clone, PartialEq, Serialize)]
-        #[serde(untagged)]
-        #[allow(missing_docs)]
-        pub enum $name {
-            $(
-                $variant(Box<$query>),
-            )*
-        }
-
-        $(
-            impl From<$query> for $name {
-                fn from(q: $query) -> Self {
-                    $name::$variant(Box::new(q))
-                }
-            }
-        )+
-
-        impl $name {
-            /// Gets aggregation name
-            pub fn name(&self) -> String {
-                match self {
-                    $(
-                        Self::$variant(a) => a.name.clone(),
-                    )+
-                }
-            }
-        }
-    };
-}
-
 #[doc(hidden)]
 #[macro_export]
 macro_rules! add_boost_and_name {
